+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.2] - 03-24-2025
+
 ### Added
 
 - The Parser class.
@@ -14,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-- cache functions.
+- Cache functions.
 
 ## [0.0.1] - 03-23-2025
 
@@ -24,5 +26,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cache functions.
 
 
-[unreleased]: https://github.com/taylorhmorris/query_and_cache/compare/v0.0.1...HEAD
+[unreleased]: https://github.com/taylorhmorris/query_and_cache/compare/v0.0.2...HEAD
+[0.0.2]: https://github.com/taylorhmorris/query_and_cache/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/taylorhmorris/query_and_cache/releases/tag/v0.0.1